requirements.txt
krakenex
pykrakenapi
python-dotenv
scipy
